backend/requirements.txt
fastapi==0.109.0
python-multipart==0.0.9
uvicorn==0.27.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
supabase==2.10.0
realtime==2.0.5
python-dotenv==1.0.0
pytest==8.4.2
pytest-asyncio==0.23.2
fastf1==3.4.0
numpy==1.26.4
pandas==2.2.0

# Voice Chat Dependencies
# Note: openai-whisper is compatible with numpy>=1.17 (our 1.26.4 is fine)
openai-whisper==20231117
pyttsx3==2.90
pydub==0.25.1
soundfile==0.12.1
aiofiles==23.2.1
# ffmpeg-python is needed for pydub audio conversion
ffmpeg-python==0.2.0
comtypes==1.4.1
pywin32==306
